+++ packages/utils/test/hashForSchema.test.ts
@@ -40,12 +40,27 @@ describe('hashForSchema', () => {
   });
 });
 
+const TEST_OBJECT_WITH_IGNORED_PROPERTIES = { a: 1, b: undefined, f: () => {} };
+
 describe('sortedJSONStringify', () => {
   it('stringifies the object with keys already in order', () => {
     expect(sortedJSONStringify(IN_ORDER_SCHEMA)).toEqual(STRINGIFIED_IN_ORDER);
   });
   it('stringifies the object putting keys in order', () => {
     expect(sortedJSONStringify(OUT_OF_ORDER_SCHEMA)).toEqual(STRINGIFIED_IN_ORDER);
+  });
+  it('stringifies objects the same as JSON.stringify', () => {
+    expect(sortedJSONStringify(TEST_OBJECT_WITH_IGNORED_PROPERTIES)).toEqual(
+      JSON.stringify(TEST_OBJECT_WITH_IGNORED_PROPERTIES),
+    );
+  });
+  it.each([
+    ['[]', []],
+    ['[undefined]', [undefined]],
+    ['[null]', [null]],
+    ['[1,undefined,2]', [1, undefined, 2]],
+  ])('stringifies %s correctly', (expected, value) => {
+    expect(sortedJSONStringify(value)).toEqual(expected);
   });
 });